# pdi/docs/openapi-pdi.yaml
openapi: 3.0.3
info:
  title: PDI API (Plataforma de Interoperabilidad)
  description: |
    API de la Plataforma de Interoperabilidad (PDI) que incluye:
    - Servicios REST para gestión de usuarios INUS (datos patronímicos)
    - Servicios REST para asociaciones usuario-prestador
    - Servicios REST de prueba y administración
    - Servicio SOAP para consulta de información de personas (DNIC)
  version: 1.0.0
  contact:
    name: PDI Development Team
servers:
  - url: http://localhost:8083
    description: Servidor local de desarrollo
paths:
...
  /api/test/personas:
    get:
      summary: Listar todas las personas
      description: Lista todas las personas en el repositorio en memoria (para testing)
      operationId: listarPersonas
      tags:
        - Test
      responses:
        '200':
          description: Lista de personas
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  personas:
                    type: array
                    items:
                      $ref: '#/components/schemas/PersonaData'
    
    post:
      summary: Agregar nueva persona
      description: Agrega una nueva persona al repositorio en memoria para pruebas
      operationId: agregarPersona
      tags:
        - Test
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonaData'
            example:
              tipoDocumento: "CI"
              numeroDocumento: "12345678"
              primerNombre: "Test"
              segundoNombre: "Usuario"
              primerApellido: "Prueba"
              segundoApellido: "Test"
              fechaNacimiento: "1990-01-01"
              sexo: "M"
              nacionalidad: "UY"
              departamento: "Montevideo"
              localidad: "Montevideo"
      responses:
        '201':
          description: Persona agregada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  mensaje:
                    type: string
                  persona:
                    $ref: '#/components/schemas/PersonaData'
                  totalPersonas:
                    type: integer
        '400':
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ============================================
  # SERVICIOS INUS - GESTIÓN DE USUARIOS
  # ============================================
...
components:
  schemas:
    PersonaData:
      type: object
      required:
        - tipoDocumento
        - numeroDocumento
        - primerNombre
        - primerApellido
      properties:
        tipoDocumento:
          type: string
          description: Tipo de documento (CI, PASAPORTE, etc.)
          example: "CI"
        numeroDocumento:
          type: string
          description: Número de documento
          example: "50830691"
        primerNombre:
          type: string
          description: Primer nombre
          example: "Victor"
        segundoNombre:
          type: string
          description: Segundo nombre
          example: "David"
        primerApellido:
          type: string
          description: Primer apellido
          example: "Alvez"
        segundoApellido:
          type: string
          description: Segundo apellido
          example: "González"
        fechaNacimiento:
          type: string
          format: date
          description: Fecha de nacimiento (YYYY-MM-DD)
          example: "2000-12-26"
        sexo:
          type: string
          enum: [M, F]
          description: Sexo (M=Masculino, F=Femenino)
          example: "M"
        nacionalidad:
          type: string
          enum: [UY, AR, BR, CL, OT]
          description: Código de nacionalidad
          example: "UY"
        departamento:
          type: string
          description: Departamento
          example: "Montevideo"
        localidad:
          type: string
          description: Localidad
          example: "Montevideo"
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        mensaje:
          type: string
          description: Mensaje de error
          example: "Usuario no encontrado con UID: uid-inexistente"
...